refactor: move confirm abort screens to `handback` module
