+++ features/handback/internal-api/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/handback/internalapi/nav/HandbackDestinations.kt
@@ -15,4 +15,7 @@ sealed interface HandbackDestinations : ProveYourIdentityDestinations {
 
     @Serializable
     data object ReturnToDesktopWeb : HandbackDestinations
+
+    @Serializable
+    data object ConfirmAbort : HandbackDestinations
 }
