+++ features/handback/internal/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/handback/internal/confirmabort/ConfirmAbort.kt
@@ -1,4 +1,4 @@
-package uk.gov.onelogin.criorchestrator.features.selectdoc.internal.abort
+package uk.gov.onelogin.criorchestrator.features.handback.internal.confirmabort
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.material3.Text
