+++ features/select-doc/internal-api/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/selectdoc/internalapi/nav/SelectDocDestinations.kt
@@ -30,7 +30,4 @@ sealed interface SelectDocDestinations : ProveYourIdentityDestinations {
 
     @Serializable
     data object ConfirmNoNonChippedID : SelectDocDestinations
-
-    @Serializable
-    data object ConfirmAbort : SelectDocDestinations
 }
