+++ features/select-doc/internal/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/selectdoc/internal/confirmnoid/nochippedid/ConfirmNoChippedIDScreen.kt
@@ -21,8 +21,8 @@ import uk.gov.android.ui.componentsv2.heading.GdsHeadingAlignment
 import uk.gov.android.ui.patterns.leftalignedscreen.LeftAlignedScreen
 import uk.gov.android.ui.theme.m3.GdsTheme
 import uk.gov.android.ui.theme.util.UnstableDesignSystemAPI
+import uk.gov.onelogin.criorchestrator.features.handback.internalapi.nav.HandbackDestinations
 import uk.gov.onelogin.criorchestrator.features.selectdoc.internal.R
-import uk.gov.onelogin.criorchestrator.features.selectdoc.internalapi.nav.SelectDocDestinations
 import uk.gov.onelogin.criorchestrator.libraries.composeutils.LightDarkBothLocalesPreview
 
 @Composable
@@ -37,7 +37,7 @@ internal fun ConfirmNoChippedIDScreen(
         viewModel.action.collect { event ->
             when (event) {
                 ConfirmNoChippedIDAction.NavigateToConfirmAbort -> {
-                    navController.navigate(SelectDocDestinations.ConfirmAbort)
+                    navController.navigate(HandbackDestinations.ConfirmAbort)
                 }
             }
         }
